ref: update buildscript, package rename
