--- .github/dependabot.yml
@@ -1,12 +1,0 @@
-version: 2
-updates:
-  - package-ecosystem: "gradle"
-    directory: "/"
-    schedule:
-      interval: "daily"
-    target-branch: "master"
-    assignees:
-      - "MrTroble"
-    ignore:
-      - dependency-name: "net.minecraftforge:forge"
-        versions: ["[1.13,]"]
