+++ gradle.properties
@@ -3,19 +3,16 @@ org.gradle.jvmargs=-Xmx3G
 org.gradle.daemon=false
 
 # Mod Meta
-mod_id=invisiblelights
 mod_name=Invisible Lights
 mod_group=com.troblecodings
-mod_author=MrTroble, Jeronimo
 mod_version=1.1.0
 
 # Minecraft
-minecraft_version=1.12.2
-loader=forge
+minecraft_version=1.14.4
 
 # Forge
-forge_version_main=14
-forge_version=14.23.5.2860
+forge_version=28.2.28
+loader=forge
 
 # Java
 java_version=8
