+++ src/main/java/com/troblecodings/invisiblelights/blocks/BlockInvisibleLight.java
@@ -1,67 +1,63 @@
 package com.troblecodings.invisiblelights.blocks;
 
-import com.troblecodings.invisiblelights.init.ILInit;
-
 import net.minecraft.block.Block;
+import net.minecraft.block.BlockRenderType;
+import net.minecraft.block.BlockState;
 import net.minecraft.block.material.Material;
-import net.minecraft.block.state.BlockFaceShape;
-import net.minecraft.block.state.IBlockState;
+import net.minecraft.client.particle.ParticleManager;
 import net.minecraft.util.BlockRenderLayer;
-import net.minecraft.util.EnumBlockRenderType;
-import net.minecraft.util.EnumFacing;
-import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.IBlockAccess;
+import net.minecraft.util.math.RayTraceResult;
+import net.minecraft.util.math.shapes.ISelectionContext;
+import net.minecraft.util.math.shapes.VoxelShape;
+import net.minecraft.util.math.shapes.VoxelShapes;
+import net.minecraft.world.IBlockReader;
+import net.minecraft.world.World;
 
 public class BlockInvisibleLight extends Block {
 
-    public BlockInvisibleLight(final int light) {
-        super(Material.GROUND);
-        this.setLightLevel(light / 15.0f);
-        this.setCreativeTab(ILInit.LIGHT_TAB);
-        this.disableStats();
-    }
+	public BlockInvisibleLight(final int light) {
+		super(Block.Properties.create(Material.EARTH)
+				.lightValue(light)
+				.hardnessAndResistance(0.5f));
+	}
 
-    @Override
-    public boolean isPassable(final IBlockAccess worldIn, final BlockPos pos) {
-        return true;
-    }
+	@Override
+	public boolean isSolid(final BlockState state) {
+		return false;
+	}
 
-    @Override
-    public boolean isOpaqueCube(final IBlockState state) {
-        return false;
-    }
+	@Override
+	public BlockRenderType getRenderType(final BlockState state) {
+		return BlockRenderType.INVISIBLE;
+	}
 
-    @Override
-    public EnumBlockRenderType getRenderType(final IBlockState state) {
-        return EnumBlockRenderType.INVISIBLE;
-    }
+	@Override
+	public BlockRenderLayer getRenderLayer() {
+		return BlockRenderLayer.CUTOUT;
+	}
 
-    @Override
-    public BlockRenderLayer getBlockLayer() {
-        return BlockRenderLayer.CUTOUT;
-    }
+	@Override
+	public VoxelShape getCollisionShape(final BlockState state, final IBlockReader worldIn, final BlockPos pos,
+			final ISelectionContext context) {
+		return VoxelShapes.empty();
+	}
 
-    @Override
-    public boolean isFullCube(final IBlockState state) {
-        return false;
-    }
+	@Override
+	public VoxelShape getShape(final BlockState state, final IBlockReader worldIn, final BlockPos pos,
+			final ISelectionContext context) {
+		return VoxelShapes.fullCube();
+	}
 
-    @Override
-    public AxisAlignedBB getBoundingBox(final IBlockState state, final IBlockAccess source, final BlockPos pos) {
-        return FULL_BLOCK_AABB;
-    }
+	@Override
+	public boolean addHitEffects(final BlockState state, final World world, final RayTraceResult target,
+			final ParticleManager manager) {
+		return true;
+	}
 
-    @Override
-    public AxisAlignedBB getCollisionBoundingBox(final IBlockState blockState, final IBlockAccess worldIn,
-            final BlockPos pos) {
-        return NULL_AABB;
-    }
-
-    @Override
-    public BlockFaceShape getBlockFaceShape(final IBlockAccess worldIn, final IBlockState state, final BlockPos pos,
-            final EnumFacing face) {
-        return BlockFaceShape.UNDEFINED;
-    }
-
+	@Override
+	public boolean addDestroyEffects(final BlockState state, final World world, final BlockPos pos,
+			final ParticleManager manager) {
+		return true;
+	}
 }
