+++ src/main/resources/META-INF/mods.toml
@@ -1,27 +1,27 @@
 modLoader="javafml"
-loaderVersion="["${forgeversion}",)"
+loaderVersion="[28,)"
 issueTrackerURL="https://github.com/MrTroble/InvisibleLights/issues"
 license="Apache License Version 2.0"
 
 [[mods]]
-modId="${modid}"
-version="${version}"
-displayName="${modname}"
+modId="invisiblelights"
+version="${file.jarVersion}"
+displayName="Invisible Lights"
 displayURL="https://troblecodings.com/"
-authors="${author}"
+authors="MrTroble, Jeronimo, Luex"
 credits="German Immersive Railroading Community"
 description="Adds invisible lights."
 
-[[dependencies.${modid}]]
+[[dependencies.invisiblelights]]
     modId="forge"
     mandatory=true
-    versionRange="["${forgeversion}",)"
+    versionRange="[28,)"
     ordering="NONE"
     side="BOTH"
 
-[[dependencies.${modid}]]
+[[dependencies.invisiblelights]]
     modId="minecraft"
     mandatory=true
-    versionRange="["${mcversion}"]"
+    versionRange="[1.14.4]"
     ordering="NONE"
     side="BOTH"
