+++ src/main/resources/pack.mcmeta
@@ -1,6 +1,6 @@
 {
     "pack": {
-        "description": "gir mod resources",
+        "description": "Invisible lights mod resources",
         "pack_format": 4
     }
 }
